Use with method instead of anonymous class
DefaultJqwikConfiguration provides the default values the configuration
properties. The default values for max tries and max discard ratio are
defined in interface PropertyDefaultValues.

Before this change the DefaultJqwikConfiguration used an anonymous
class of the PropertyDefaultValues interface in order to provide the
default values. However, the PropertyDefaultValues interface already has
a smart constructor / factory method with to create an instance with the
given values.

This commit replaces the anonymous class with the use of with. Thus,
there is less code duplication and better readability.

I hereby agree to the terms of the jqwik Contributor Agreement.