Moved uniqe id comparison to TestRun
Class TestRunData provides method byUniqueId to find a test run by the a
UniqueId.

Before this commit, this method reached into the structure of TestRun in
order to compare unique ids. This results into a higher coupling.

This commit moves the part of comparing uniqe ids into TestRun. The new
query method is named hasUniqueId and compares given id with the id of
this test run.