Simplified Optional chain
Method determineSeed determines the seed for the random generator for a
given property. If a seed is defined by the Property annotation, it is
used over the stored seed from a previous run.

Before this commit, the method used the trinary operator in a call to
map. The orElse method would never be called.

This commit simplifies this call by using filter and orElse instead.