Simplified Optional chain
Method determineSeed determines the seed for the random generator for a
given property. If a seed is defined by the Property annotation, it is
used over the stored seed from a previous run.

Before this commit, the method used a filter in order to determine the
seed. However, the filter condition is independent from the seed of the
failed run and can be hoisted to the top of the method.