Ignore Group annotation on static inner classes
The TopLevelContainerDiscoverSpec determines if a class is a container
on top level or not. A class is on top level if it is the top most
public class or if it is a static inner class.

Before this commit, the check for static classes was not done properly.
Instead of checking the candidate, the Group annotation was checked.

This commit replaces Group.class by candidate and, thus, the check is
done properly. A test was added with a single nested container annotated
with Group. Within the discovery phase, the group annotation should be
ignored like it is done with other top level containers.