Improve FeatureExtractor: use contains for identity extractors, and avoid building intermediate sets
Previously, FeatureExtractor built an itermediate set for a sole purpose
of uniqueness check.

The change delegates to .contains(value) for the common case of identity extractor,
and we use noneMatch to avoid building the set for non-default extractors.

It might be that ContainerGenerator requires improvements to keep per-feature
sets, however, the current fix is a low-hanging fruit.

fixes #333