test: add IBM Semeru Java distribution to test matrix
Semeru uses Open9J JIT, so it might behave differently from the other OpenJDK
distributions, so it is assigned a greater weight in the test matrix.

The option for testing with "same hashcode" -XX:hashCode=2 is moved
to Test task execution only since javac, and kotlinc do not behave
well with "same hashcode".